debug /getdata

+++ app/models.py
@@ -1,9 +1,9 @@
 from pydantic import BaseModel
 
 class SensorData(BaseModel):
-    temperature: int
-    humidity: int
-    moisture: int
+    temperature: float
+    humidity: float
+    moisture: float
 
 class TaskCreate(BaseModel):
     title: str
